Rearrange some dicts file
