+++ TextRecognitionDataGenerator/fonts/jp/font_dict.pkl
