+++ TextRecognitionDataGenerator/texts/full_format_db.date
