+++ run_examples.txt
@@ -4,4 +4,4 @@ python3 run.py -l en -wk -pre sjnk -na 3 -rk -k 0.5 -rbl -bl 2 -b 0 -f 30 -t 5 -
 
 python3 run.py -l sjnk -sjnk -pre space -na 3 -rk -k 0.5 -rbl -bl 2 -b 0 -f 30 -t 4 -c 20000
 
-python run.py -l jp -i full_format_db.date -pre jp -na 3 -rk -k 0.4 -rbl -bl 1 -b 1 -f 30 -t 4 -c 400 -w 10
+python run.py -l jp -i ./texts/full_format_db.date -pre jp -na 3 -rk -k 0.4 -rbl -bl 1 -b 1 -f 30 -t 4 -c 400 -w 10
